pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "anti-standupbot"
version = "0.1.0"
description = "Zero-fluff async standup logger with Claude AI weekly digests"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
standup = "standup.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["standup"]
